Add files via upload
I develop a web application that allows users to submit ratings for stores registered on the platform.
The ratings should range from 1 to 5.
A single login system should be implemented for all users. Based on their roles, users will have access to different
functionalities upon logging in. Normal users should be able to sign up on the platform through a registration page.